shorten pipex function
