+++ includes/pipex.h
@@ -6,7 +6,7 @@
 /*   By: gyong-si <gyong-si@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/12 10:22:41 by gyong-si          #+#    #+#             */
-/*   Updated: 2024/04/23 14:58:54 by gyong-si         ###   ########.fr       */
+/*   Updated: 2024/04/24 13:08:56 by gyong-si         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,6 +20,7 @@
 # include <sys/types.h>
 # include <sys/wait.h>
 # include <unistd.h>
+# include "tokenizer.h"
 
 int		open_file(char *file, int n);
 char	*my_getenv(char *name, char **env);
@@ -27,6 +28,7 @@ char	*get_path(char *cmd, char **env);
 void	exec_cmd(char *cmd, char **env);
 void	ft_free_tab(char **tab);
 void	exit_handler(int exit_code);
+void	pipex(t_token *token_lst, char **env);
 void	child(char *command, int *p_fd, char **env);
 void 	parent(int *p_fd);
 void	do_pipe(char *command, char **env);
