+++ src/pipex/pipex.c
@@ -6,7 +6,7 @@
 /*   By: gyong-si <gyong-si@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/12 10:24:38 by gyong-si          #+#    #+#             */
-/*   Updated: 2024/04/23 22:50:37 by gyong-si         ###   ########.fr       */
+/*   Updated: 2024/04/24 13:51:41 by gyong-si         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,6 +19,75 @@ eroor messages, free the allocated memmory and then exits the program */
 //for (int i = 0; s_cmd[i] != NULL; i++) {
 //	printf("s_cmd[%d]: %s\n", i, s_cmd[i]);
 //}
+
+/***
+void	pipex(char *input, char **env)
+{
+	int		n;
+	int		i;
+	char	**command;
+
+	command = ft_dqsplit(input, ' ');
+	convert_cmd(command);
+	i = 0;
+	n = 0;
+	while (command[n] != NULL)
+		n++;
+	while (i < n - 1)
+	{
+		do_pipe(command[i++], env);
+	}
+	exec_cmd(command[i], env);
+} **/
+
+static char	**init_command(t_token *token_lst, int num_of_command)
+{
+	t_token	*curr_token;
+	char	**command;
+	int	i;
+
+	command = (char **)malloc((num_of_command + 1) * sizeof(char *));
+	if (!command)
+		exit(EXIT_FAILURE);
+	i = 0;
+	curr_token = token_lst;
+	while (i < num_of_command)
+	{
+		if (curr_token->type == T_IDENTIFIER)
+		{
+			command[i++] = curr_token->token;
+		 	printf("command[%d]: %s\n", i, curr_token->token);
+		}
+		curr_token = curr_token->next;
+	}
+	command[num_of_command] = NULL;
+	return (command);
+}
+
+void	pipex(t_token *token_lst, char **env)
+{
+	int		i;
+	int		num_of_command;
+	t_token	*curr_token;
+	char	**command;
+
+	curr_token = token_lst;
+	num_of_command = 0;
+	while (curr_token != NULL)
+	{
+		if (curr_token->type == T_IDENTIFIER)
+			num_of_command++;
+		curr_token = curr_token->next;
+	}
+	command = init_command(token_lst, num_of_command);
+	i = 0;
+	while (i < num_of_command - 1)
+	{
+		do_pipe(command[i++], env);
+	}
+	exec_cmd(command[num_of_command - 1], env);
+	free(command);
+}
 
 void	exec_cmd(char *cmd, char **env)
 {
